Move submit logic to LoginModel::submit_login_form
