+++ src/app/components/login/login_model.rs
@@ -1,3 +1,5 @@
+use gtk::{EntryExt, WidgetExt};
+
 use crate::app::credentials;
 use crate::app::{ActionDispatcher, AppAction};
 
@@ -30,5 +32,17 @@ impl LoginModel {
 
     pub fn login(&self, u: String, p: String) {
         self.dispatcher.dispatch(AppAction::TryLogin(u, p));
+    }
+
+    pub fn submit_login_form(&self, username: gtk::Entry, password: gtk::Entry) {
+        let username_text = username.get_text().as_str().to_string();
+        let password_text = password.get_text().as_str().to_string();
+        if username_text.is_empty() {
+            username.grab_focus();
+        } else if password_text.is_empty() {
+            password.grab_focus();
+        } else {
+            self.login(username_text, password_text);
+        }
     }
 }
